cli(node): process node flags
